LayoutTests/accessibility/aria-namefrom-author-expected.txt

This tests all the cases where nameFrom: author is used instead of nameFrom: contents. This means that if these elements are used in aria-labelledby they should not return their inner text. The button should retain its aria-label.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS buttonAccName == button.getAttribute('aria-label') is true
PASS buttonAccName != button.innerText is true
PASS successfullyParsed is true

TEST COMPLETE
